requirements.txt
langchain
langchain-community
langchain-openai

pypdf
python-dotenv

ragas

datasets
tqdm
chromadb